appops: Disassociate appops requests from notification broadcasts
The AppOpsService is essentially a manager for a set of counters
and permissions. Each operation request has the potential to change
the state and, as such, access to such state is synchronized.

We are whitnessing deadlocks caused by the broadcast and, in fact,
while we want to notify superuser changes eventually, it does not
have to be synchronous with the app ops request. This patch uses the
request to schedule the notification on a handler, leaving the locking
semantics of appops intact.

Change-Id: I94f6dd2c66b9492f95d3c9ffb438b3e6417007d7

Conflicts:
	services/core/java/com/android/server/AppOpsService.java